Custom frontmatter handling solution
Introducing custom frontmatter handling solution for smoother experience. It instantly updates the "modified" timestamp value with each keystroke, and doesn't force any YAML convention on the rest of your frontmatter. The only downside is that it might be having trouble with nested keys.

Custom frontmatter handling solution

Introducing custom frontmatter handling solution for smoother experience. It instantly updates the "modified" timestamp value with each keystroke, and doesn't force any YAML convention on the rest of your frontmatter. The only downside is that it might be having trouble with nested keys.

Custom frontmatter handling solution

Introducing custom frontmatter handling solution for smoother experience. It instantly updates the "modified" timestamp value with each keystroke, and doesn't force any YAML convention on the rest of your frontmatter. The only downside is that it might be having trouble with nested keys.